dash/requirements.txt
flask
requests
numpy
Pillow
safetensors
torch
torchvision
transformers>=4.56.0
